feat(governance): ts, split components
